# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: BehaveDrift API
  version: 1.0.0
  description: |
    **BehaveDrift API** is an open-source AI inference engine that detects
    behavioural pattern drift in dementia residents.

    It accepts longitudinal care observation data and returns:
    - A **drift score** quantifying deviation from each resident's personal baseline
    - **Tiered alerts** (T1–T4) with natural-language explanations
    - **FHIR R4** Condition resources for clinical handover

    ### Authentication

    Two modes are supported:

    **OAuth2 Client Credentials (recommended for production):**
    ```
    POST /v1/auth/token
    grant_type=client_credentials&client_id=...&client_secret=...
    Authorization: Bearer <jwt>
    ```

    **API Key (for self-hosted / development):**
    ```
    X-API-Key: bda_sk_xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
    ```

    ### Error Responses
    All errors use a consistent shape:
    ```json
    {
      "error": "validation_error",
      "message": "Field 'resident_id' is required.",
      "code": 422,
      "details": [],
      "request_id": "req_abc123"
    }
    ```

    ### Versioning
    All endpoints are prefixed with `/v1/`. Breaking changes increment to `/v2/`.

  contact:
    name: Umar Isa
    email: hello@behavedrift.io
    url: https://github.com/umarisa/behavedrift-api

  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OAuth2 Client Credentials JWT (RS256, 60 min TTL)

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key for self-hosted / development use

  # ============================================================
  # SHARED SCHEMAS
  # ============================================================
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, message, code, request_id]
      properties:
        error:
          type: string
          description: Machine-readable error code (snake_case)
          example: validation_error
        message:
          type: string
          description: Human-readable explanation
          example: "Field 'resident_id' is required and cannot be blank."
        code:
          type: integer
          description: HTTP status code
          example: 422
        details:
          type: array
          description: Field-level validation errors (422 only)
          items:
            type: object
            properties:
              field:
                type: string
                example: resident_id
              issue:
                type: string
                example: required
              value:
                nullable: true
                example: null
        request_id:
          type: string
          description: Unique request ID for log correlation
          example: req_4f3a9c21d
        documentation_url:
          type: string
          format: uri
          example: https://docs.behavedrift.io/errors/validation_error

    # ---------- Pagination ----------
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 142
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
        has_next:
          type: boolean
          example: true

    # ---------- Resident ----------
    ResidentCreate:
      type: object
      required: [internal_reference]
      properties:
        internal_reference:
          type: string
          description: Your platform's internal ID for this resident (not stored as PII)
          example: RES-001
        date_of_birth:
          type: string
          format: date
          description: Used only for age-band analytics — never stored in logs
          example: "1938-04-15"
        diagnosis_codes:
          type: array
          description: ICD-10 or SNOMED codes (e.g. F00.1 = dementia in Alzheimer's)
          items:
            type: string
          example: ["F00.1"]
        baseline_window_days:
          type: integer
          minimum: 14
          maximum: 90
          default: 28
          description: Rolling window (days) for baseline calculation
          example: 28
        risk_profile:
          type: string
          enum: [low, medium, high]
          default: medium
          description: Clinical risk profile — affects alert sensitivity
          example: high
        notes:
          type: string
          maxLength: 500
          example: "History of recurrent UTIs. Baseline resets post-hospitalisation."
    ResidentResponse:
      type: object
      properties:
        resident_id:
          type: string
          example: res_7f3a9c21
        internal_reference:
          type: string
          example: RES-001
        baseline_status:
          type: string
          enum: [initialising, active, reset_pending, suspended]
          example: active
        baseline_window_days:
          type: integer
          example: 28
        risk_profile:
          type: string
          enum: [low, medium, high]
          example: high
        diagnosis_codes:
          type: array
          items:
            type: string
          example: ["F00.1"]
        total_observations:
          type: integer
          example: 187
        min_observations_required:
          type: integer
          example: 14
        created_at:
          type: string
          format: date-time
          example: "2026-01-10T09:00:00Z"
        last_observation_at:
          type: string
          format: date-time
          example: "2026-03-14T14:30:00Z"
    BaselineSummary:
      type: object
      properties:
        resident_id:
          type: string
          example: res_7f3a9c21
        baseline_status:
          type: string
          enum: [initialising, active, reset_pending, suspended]
        window_days:
          type: integer
          example: 28
        window_start:
          type: string
          format: date-time
        window_end:
          type: string
          format: date-time
        total_observations_in_window:
          type: integer
          example: 84
        signals:
          type: object
          description: Per-signal baseline statistics
          additionalProperties:
            type: object
            properties:
              mean:
                type: number
              std_dev:
                type: number
              most_common:
                type: string
              sample_count:
                type: integer
          example:
            mood:
              mean: 3.7
              std_dev: 0.6
              sample_count: 82
            appetite:
              most_common: good
              sample_count: 79
        last_calculated_at:
          type: string
          format: date-time

    # ---------- Signals ----------
...
    WebhookCreate:
      type: object
      required: [url, events]
      properties:
        url:
          type: string
          format: uri
          example: https://your-platform.com/webhooks/behavedrift
        events:
          type: array
          minItems: 1
          items:
            type: string
            enum:
              - alert.created
              - alert.tier_escalated
              - alert.acknowledged
              - alert.resolved
              - baseline.updated
              - resident.risk_status_changed
          example: [alert.created, alert.tier_escalated]
        description:
          type: string
          example: Nourish integration endpoint
        active:
          type: boolean
          default: true
        min_tier:
          type: string
          enum: [T1, T2, T3, T4]
          default: T1
          description: Only send alerts at or above this tier
    WebhookResponse:
      type: object
      properties:
        webhook_id:
          type: string
          example: wh_a1b2c3d4
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
        active:
          type: boolean
        signing_secret:
          type: string
          description: Returned only on creation — store securely, not shown again
          example: whsec_xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
        created_at:
          type: string
          format: date-time
        last_triggered_at:
          type: string
          format: date-time
          nullable: true

    # ---------- Auth ----------
...
  responses:
    Unauthorized:
      description: Authentication required or token expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: authentication_required
            message: A valid Bearer token or X-API-Key header is required.
            code: 401
            request_id: req_abc123
    Forbidden:
      description: Authenticated but insufficient scope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: forbidden
            message: Your token does not have the required scope.
            code: 403
            request_id: req_abc123
    NotFound:
      description: Resource not found (or belongs to another tenant)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: not_found
            message: The requested resource does not exist.
            code: 404
            request_id: req_abc123
    UnprocessableEntity:
      description: Request body failed schema validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: validation_error
            message: "Field 'resident_id' is required."
            code: 422
            details:
              - field: resident_id
                issue: required
                value: null
            request_id: req_abc123
    RateLimited:
      description: Too many requests
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: rate_limit_exceeded
            message: You have exceeded 300 requests per minute. Retry after 30 seconds.
            code: 429
            request_id: req_abc123
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: internal_error
            message: An unexpected error occurred. The request_id has been logged.
            code: 500
            request_id: req_abc123

# ============================================================
# SECURITY (applied globally by default)
# ============================================================
security:
  - BearerAuth: []
  - ApiKeyAuth: []

# ============================================================
# PATHS
# ============================================================
paths:
...
  /v1/residents:
    post:
      tags: [Residents]
      summary: Register a resident
      description: Registers a new resident and initialises their baseline profile.
      operationId: createResident
      security:
        - BearerAuth: [behavedrift:write]
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResidentCreate'
      responses:
        '201':
          description: Resident registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResidentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: internal_reference already registered for this tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: conflict
                message: A resident with internal_reference 'RES-001' already exists.
                code: 409
                request_id: req_abc123
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags: [Residents]
      summary: List residents
      description: Returns a paginated list of all residents registered for the tenant.
      operationId: listResidents
      security:
        - BearerAuth: [behavedrift:read]
        - ApiKeyAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1, minimum: 1 }
        - name: page_size
          in: query
          schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
        - name: baseline_status
          in: query
          schema:
            type: string
            enum: [initialising, active, reset_pending, suspended]
        - name: risk_profile
          in: query
          schema:
            type: string
            enum: [low, medium, high]
      responses:
        '200':
          description: Paginated list of residents
          content:
            application/json:
              schema:
                type: object
                properties:
                  residents:
                    type: array
                    items:
                      $ref: '#/components/schemas/ResidentResponse'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/residents/{resident_id}:
    parameters:
      - name: resident_id
        in: path
        required: true
        schema:
          type: string
        example: res_7f3a9c21

    get:
      tags: [Residents]
      summary: Get a resident
      operationId: getResident
      security:
        - BearerAuth: [behavedrift:read]
        - ApiKeyAuth: []
      responses:
        '200':
          description: Resident details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResidentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Residents]
      summary: Delete a resident (GDPR erasure)
      description: |
        Permanently deletes all resident data. This satisfies UK GDPR Article 17.
        A verifiable deletion certificate is returned.
        **This action is irreversible.**
      operationId: deleteResident
      security:
        - BearerAuth: [behavedrift:admin]
        - ApiKeyAuth: []
      responses:
        '200':
          description: Resident data erased
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted_resident_id: { type: string }
                  certificate_id: { type: string }
                  deleted_at: { type: string, format: date-time }
                  records_deleted:
                    type: object
                    properties:
                      observations: { type: integer }
                      alerts: { type: integer }
                      baseline_snapshots: { type: integer }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/residents/{resident_id}/baseline:
    parameters:
      - name: resident_id
        in: path
        required: true
        schema:
          type: string

    get:
      tags: [Residents]
      summary: Get resident baseline
      description: Returns the current baseline model for a resident.
      operationId: getBaseline
      security:
        - BearerAuth: [behavedrift:read]
        - ApiKeyAuth: []
      responses:
        '200':
          description: Baseline model summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaselineSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/residents/{resident_id}/baseline/reset:
    parameters:
      - name: resident_id
        in: path
        required: true
        schema:
          type: string

    put:
      tags: [Residents]
      summary: Reset resident baseline
      description: |
        Discards the current baseline and starts a fresh accumulation window.
        Use this after a hospitalisation, significant care plan change, or medication review.
      operationId: resetBaseline
      security:
        - BearerAuth: [behavedrift:write]
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  enum:
                    - post_hospitalisation
                    - medication_change
                    - care_plan_change
                    - clinical_review
                    - other
                  example: post_hospitalisation
                notes:
                  type: string
                  maxLength: 300
                  example: Returned from 5-day hospital stay. New medication prescribed.
      responses:
        '200':
          description: Baseline reset initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  resident_id: { type: string }
                  baseline_status:
                    type: string
                    example: initialising
                  reset_at: { type: string, format: date-time }
                  min_observations_required: { type: integer, example: 14 }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---- Observations ----
...
  /v1/webhooks:
    post:
      tags: [Webhooks]
      summary: Register a webhook
      description: |
        Register an endpoint to receive BehaveDrift events.
        All events are signed with HMAC-SHA256 using the returned `signing_secret`.
        **The signing_secret is shown only once — store it immediately.**
      operationId: createWebhook
      security:
        - BearerAuth: [behavedrift:admin]
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookCreate'
      responses:
        '201':
          description: Webhook registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

    get:
      tags: [Webhooks]
      summary: List registered webhooks
      operationId: listWebhooks
      security:
        - BearerAuth: [behavedrift:admin]
        - ApiKeyAuth: []
      responses:
        '200':
          description: List of registered webhooks
          content:
            application/json:
              schema:
                type: object
                properties:
                  webhooks:
                    type: array
                    items:
                      $ref: '#/components/schemas/WebhookResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/webhooks/{webhook_id}:
    delete:
      tags: [Webhooks]
      summary: Remove a webhook
      operationId: deleteWebhook
      security:
        - BearerAuth: [behavedrift:admin]
        - ApiKeyAuth: []
      parameters:
        - name: webhook_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Webhook removed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'